chore: extract `vite` config to its own file  - create `vite.config.js`
