+++ packages/core/.storybook/main.js
@@ -1,8 +1,4 @@
-import path from 'path';
 import { readdirSync } from 'fs';
-import { mergeConfig } from 'vite';
-import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
-import { ImageLoader } from 'esbuild-vanilla-image-loader';
 
 export default {
   staticDirs: ['../components/GlobalStyle'],
@@ -32,33 +28,5 @@ export default {
   },
   docs: {
     autodocs: 'tag',
-  },
-  viteFinal: async function viteFinal(config) {
-    return mergeConfig(config, {
-      build: { chunkSizeWarningLimit: 1600 },
-      plugins: [
-        vanillaExtractPlugin({
-          identifiers: ({ filePath, hash }) => {
-            if (
-              filePath.startsWith('components/themes') &&
-              !filePath.endsWith('contract.css.ts')
-            ) {
-              const file = path.basename(filePath, '.css.ts');
-
-              return `r95_theme_${file}_${hash}`;
-            }
-
-            return `r95_${hash}`;
-          },
-          esbuildOptions: {
-            plugins: [
-              ImageLoader({
-                filter: /\.(png|svg|ttf|eot|woff|woff2)$/,
-              }),
-            ],
-          },
-        }),
-      ],
-    });
   },
 };
